# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Чат для Системы HRM",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/files/{id}/download": {
            "get": {
                "security": [
                    {
                        "LoginHeaderAuth": []
                    }
                ],
                "description": "Скачивает прикрепленный к сообщению файл по его идентификатору (fileID)",
                "tags": [
                    "файлы"
                ],
                "summary": "Скачать прикрепленный файл",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID файла",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "LoginHeaderAuth": {
            "description": "Передайте имя пользователя или email для авторизации (например: amin, fed, ernest или far@gmail.com)",
            "type": "apiKey",
            "name": "Login",
            "in": "header"
        }
    }
}